hello

+++ solve.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 void solve()
 {
-    long long a,b;
+    int a,b;
     cin>>a>>b;
     cout<<a*b<<endl;
 }
